requirements.txt
nuscenes-devkit
ultralytics
mlx-vlm
mlx
torch
torchvision
opencv-python
matplotlib
Pillow